feat: 숫자의 길이와 중복을 검증하는 method 추가
